fix: resolve 8 review findings across security, portability, and display
CRITICAL fixes:
- BSGS_BUFFERXPOINTLENGTH 0->16: bloom filters were hashing zero bytes,
  making all checks return true and defeating the 3-tier filter pipeline
- cycles/j uint32_t->uint64_t in 5 BSGS thread functions: prevented
  truncation that silently skipped search ranges above ~4.4T entries
- bsgs_secondcheck/thirdcheck parameter a widened to uint64_t: prevents
  wrong key reconstruction when j*1024+i exceeds 32-bit range
- AVX2 ModMulK1 carry accumulation: changed 0 to t[4] at reduction step
  to preserve accumulated carries from prior partial products

HIGH fixes:
- Shell injection: validate g_executable_path and g_target_file with
  is_safe_path() before popen() in wizard_client.c
- Thread safety: localtime() replaced with localtime_r() in benchmark.cpp
- bsgs_thirdcheck negation special case now verifies with ComputePublicKey
  before reporting key found
- n_range_end underflow protection: guard Sub with IsGreaterOrEqual check
  in backward/dance/both BSGS modes

+++ src/search/search_context.h
@@ -303,8 +303,8 @@ void *thread_bPload_2blooms(void *vargp);
 /*  Sort/search functions declared in bsgs/bsgs_sort.h (extern "C")   */
 /* ------------------------------------------------------------------ */
 
-int bsgs_secondcheck(Int *start_range, uint32_t a, uint32_t k_index, Int *privatekey);
-int bsgs_thirdcheck(Int *start_range, uint32_t a, uint32_t k_index, Int *privatekey);
+int bsgs_secondcheck(Int *start_range, uint64_t a, uint32_t k_index, Int *privatekey);
+int bsgs_thirdcheck(Int *start_range, uint64_t a, uint32_t k_index, Int *privatekey);
 void calcualteindex(int i, Int *key);
 
 #endif /* SEARCH_CONTEXT_H */
